+++ gnark/gnark-jni/gnark-eip-196.go
@@ -17,6 +17,11 @@ import (
     "github.com/consensys/gnark-crypto/ecc/bn254/fp"
 )
 
+var ErrMalformedPointEIP196 = errors.New("invalid point encoding")
+var ErrInvalidInputPairingLengthEIP196 = errors.New("invalid input parameters, invalid input length for pairing")
+var ErrPointNotInFieldEIP196 = errors.New("point not in field")
+var ErrPointOnCurveCheckFailedEIP196 = errors.New("point is not on curve")
+
 const (
     EIP196PreallocateForResult = 128
     EIP196PreallocateForError = 256
@@ -26,81 +31,88 @@ const (
 	EIP196PreallocateForG2 = EIP196PreallocateForG1 * 2 // G2 points are encoded as 2 concatenated G1 points
 )
 
-var ErrSubgroupCheckFailed = errors.New("invalid point: subgroup check failed")
-var ErrPointOnCurveCheckFailed = errors.New("invalid point: point is not on curve")
-var ErrMalformedOutputBytes = errors.New("malformed output buffer parameter")
+
+// bn254Modulus is the value 21888242871839275222246405745257275088696311157297823662689037894645226208583
+var bn254Modulus = new(big.Int).SetBytes([]byte{
+	0x30, 0x64, 0x4e, 0x72, 0xe1, 0x31, 0xa0, 0x29,
+	0xb8, 0x50, 0x45, 0xb6, 0x81, 0x81, 0x58, 0x5d,
+	0x97, 0x81, 0x6a, 0x91, 0x68, 0x71, 0xca, 0x8d,
+	0x3c, 0x20, 0x8c, 0x16, 0xd8, 0x7c, 0xfd, 0x47,
+})
 
 // Predefine a zero slice of length 16
-var zeroSlice = make([]byte, 16)
+var zeroEIP196Slice = make([]byte, 16)
 
 //export eip196altbn128G1Add
-func eip196altbn128G1Add(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen, cOutputLen, cErrorLen C.int) C.int {
+func eip196altbn128G1Add(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen C.int, cOutputLen, cErrorLen *C.int) C.int {
     inputLen := int(cInputLen)
-    errorLen := int(cErrorLen)
-
-
-    if (inputLen > 2*EIP196PreallocateForG1) {
-      // trunc if input too long
-      inputLen = 2*EIP196PreallocateForG1
-    }
-
-    if inputLen < EIP196PreallocateForG1 {
-        // if we do not have complete input, return 0
-        return 0
-    }
+    errorLen := (*int)(unsafe.Pointer(cErrorLen))
+    outputLen := (*int)(unsafe.Pointer(cOutputLen))
 
     // Convert error C pointers to Go slices
-    errorBuf := castErrorBuffer(javaErrorBuf, errorLen)
+    errorBuf := castErrorBufferEIP196(javaErrorBuf, errorLen)
 
+    if (inputLen > 2*EIP196PreallocateForG1) {
+        // trunc if input too long
+        inputLen = 2*EIP196PreallocateForG1
+    }
 
     // Convert input C pointers to Go slices
     input := (*[2*EIP196PreallocateForG1]byte)(unsafe.Pointer(javaInputBuf))[:inputLen:inputLen]
 
+    if (inputLen == 0) {
+        *outputLen = 0
+        return 0
+    }
+
     // generate p0 g1 affine
-    var p0 bn254.G1Affine
-    err := p0.Unmarshal(input[:64])
+    p0, err := safeUnmarshalEIP196(input, 0)
+
     if err != nil {
-        copy(errorBuf, err.Error())
+        dryError(err, errorBuf, outputLen, errorLen)
         return 1
     }
 
     if inputLen < 2*EIP196PreallocateForG1 {
         // if incomplete input is all zero, return p0
-        if isAllZero(input[64:inputLen]) {
+        if isAllZeroEIP196(input, 64) {
             ret := p0.Marshal()
             g1AffineEncode(ret, javaOutputBuf)
+            *outputLen = EIP196PreallocateForG1
+            return 0;
         }
-        return 0;
     }
-
     // generate p1 g1 affine
-    var p1 bn254.G1Affine
-    err = p1.Unmarshal(input[64:])
+    p1, err := safeUnmarshalEIP196(input, 64)
 
     if err != nil {
-        copy(errorBuf, err.Error())
+        dryError(err, errorBuf, outputLen, errorLen)
         return 1
     }
+
     // Use the Add method to combine points
-    result := p0.Add(&p0, &p1)
+    result := p0.Add(p0, p1)
 
     // marshal the resulting point and encode directly to the output buffer
     ret := result.Marshal()
     g1AffineEncode(ret, javaOutputBuf)
+    *outputLen = EIP196PreallocateForG1
     return 0
 
 }
 
 //export eip196altbn128G1Mul
-func eip196altbn128G1Mul(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen, cOutputLen, cErrorLen C.int) C.int {
+func eip196altbn128G1Mul(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen C.int, cOutputLen, cErrorLen *C.int) C.int {
     inputLen := int(cInputLen)
-    errorLen := int(cErrorLen)
+    errorLen := (*int)(unsafe.Pointer(cErrorLen))
+    outputLen := (*int)(unsafe.Pointer(cOutputLen))
 
     // Convert error C pointers to Go slices
-    errorBuf := castErrorBuffer(javaErrorBuf, errorLen)
+    errorBuf := castErrorBufferEIP196(javaErrorBuf, errorLen)
 
     if inputLen < EIP196PreallocateForG1 {
         // if we do not have complete input, return 0
+        *outputLen = 0
         return 0
     }
 
@@ -113,11 +125,10 @@ func eip196altbn128G1Mul(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInp
     input := (*[EIP196PreallocateForG1 + EIP196PreallocateForScalar]byte)(unsafe.Pointer(javaInputBuf))[:inputLen:inputLen]
 
     // generate p0 g1 affine
-    var p0 bn254.G1Affine
-    err := p0.Unmarshal(input[:64])
+    p0, err := safeUnmarshalEIP196(input, 0)
 
     if err != nil {
-        copy(errorBuf, err.Error())
+        dryError(err, errorBuf, outputLen, errorLen)
         return 1
     }
 
@@ -133,25 +144,28 @@ func eip196altbn128G1Mul(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInp
     scalar.SetBytes(scalarBytes[:])
 
     // multiply g1 point by scalar
-    result := p0.ScalarMultiplication(&p0, scalar)
+    result := p0.ScalarMultiplication(p0, scalar)
 
     // marshal the resulting point and encode directly to the output buffer
     ret := result.Marshal()
     g1AffineEncode(ret, javaOutputBuf)
+    *outputLen = EIP196PreallocateForG1
     return 0
 }
 
 //export eip196altbn128Pairing
-func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen, cOutputLen, cErrorLen C.int) C.int {
+func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cInputLen C.int, cOutputLen, cErrorLen *C.int) C.int {
     inputLen := int(cInputLen)
-    outputLen := int(cOutputLen)
-    errorLen := int(cErrorLen)
+    errorLen := (*int)(unsafe.Pointer(cErrorLen))
+    outputLen := (*int)(unsafe.Pointer(cOutputLen))
 
     // Convert error C pointers to Go slices
-    output := castBuffer(javaOutputBuf, outputLen)
+    output := castBufferEIP196(javaOutputBuf, outputLen)
 
     // Convert error C pointers to Go slices
-    errorBuf := castErrorBuffer(javaErrorBuf, errorLen)
+    errorBuf := castErrorBufferEIP196(javaErrorBuf, errorLen)
+
+    *outputLen = 32
 
     if inputLen == 0 {
         output[31]=0x01
@@ -159,12 +173,12 @@ func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cI
     }
 
     if inputLen % (EIP196PreallocateForG2 + EIP196PreallocateForG1) != 0 {
-        copy(errorBuf, "invalid input parameters, invalid input length for pairing\x00")
+        dryError(ErrInvalidInputPairingLengthEIP196, errorBuf, outputLen, errorLen)
         return 1
     }
 
     // Convert input C pointers to Go slice
-    input := castBufferToSlice(unsafe.Pointer(javaInputBuf), inputLen)
+    input := castBufferToSliceEIP196(unsafe.Pointer(javaInputBuf), inputLen)
 
     var pairCount = inputLen / (EIP196PreallocateForG2 + EIP196PreallocateForG1)
     g1Points := make([]bn254.G1Affine, pairCount)
@@ -178,7 +192,7 @@ func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cI
         err := g1.Unmarshal(input[i*192:i*192+64])
 
         if err != nil {
-            copy(errorBuf, err.Error())
+            dryError(err, errorBuf, outputLen, errorLen)
             return 1
         }
 
@@ -187,7 +201,7 @@ func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cI
         err = g2.Unmarshal(input[i*192+64:(i+1)*192])
 
         if err != nil {
-            copy(errorBuf, err.Error())
+            dryError(err, errorBuf, outputLen, errorLen)
             return 1
         }
 
@@ -198,7 +212,7 @@ func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cI
 
     isOne, err := bn254.PairingCheck(g1Points, g2Points)
     if err != nil {
-        copy(errorBuf, err.Error())
+        dryError(err, errorBuf, outputLen, errorLen)
         return -1
     }
 
@@ -206,7 +220,6 @@ func eip196altbn128Pairing(javaInputBuf, javaOutputBuf, javaErrorBuf *C.char, cI
         // respond with 1 if pairing check was true, leave 0's intact otherwise
         output[31]=0x01
     }
-
     return 0
 
 }
@@ -226,7 +239,69 @@ func g1AffineEncode(g1Point []byte, output *C.char) (error) {
     return nil
 }
 
-func castBufferToSlice(buf unsafe.Pointer, length int) []byte {
+func safeUnmarshalEIP196(input []byte, offset int) (*bn254.G1Affine, error) {
+    var g1 bn254.G1Affine
+
+    if (len(input) - offset < 64) {
+        return nil, ErrMalformedPointEIP196
+    }
+
+    if !checkInFieldEIP196(input[offset: offset + 32]) {
+        return nil, ErrPointNotInFieldEIP196
+    }
+
+    err := g1.X.SetBytesCanonical(input[offset:offset + 32])
+
+    if (err == nil) {
+
+        if !checkInFieldEIP196(input[offset + 32: offset + 64]) {
+            return nil, ErrPointNotInFieldEIP196
+        }
+        err := g1.Y.SetBytesCanonical(input[offset + 32:offset + 64])
+        if (err == nil) {
+            if (!g1.IsOnCurve()) {
+                return nil, ErrPointOnCurveCheckFailedEIP196
+            }
+            return &g1, nil
+        }
+    }
+
+
+    return nil, err
+}
+
+// checkInField checks that an element is in the field, not-in-field will normally
+// be caught during unmarshal, but here in case of no-op calls of a single parameter
+func checkInFieldEIP196(data []byte) bool {
+
+	// Convert the byte slice to a big.Int
+	elem := new(big.Int).SetBytes(data)
+
+	// Compare the value to the bn254Modulus
+	return bn254Modulus.Cmp(elem) == 1
+}
+
+// isAllZero checks if all elements in the byte slice are zero
+func isAllZeroEIP196(data []byte, offset int) bool {
+    if len(data) > 64 {
+        slice := data [offset:]
+        for _, b := range slice {
+            if b != 0 {
+                return false
+            }
+        }
+    }
+    return true
+}
+
+func dryError(err error, errorBuf []byte, outputLen, errorLen *int) {
+    errStr := "invalid input parameters, " + err.Error();
+    copy(errorBuf, errStr)
+    *outputLen = 0
+    *errorLen = len(errStr)
+}
+
+func castBufferToSliceEIP196(buf unsafe.Pointer, length int) []byte {
     var slice []byte
     // Obtain the slice header
     header := (*reflect.SliceHeader)(unsafe.Pointer(&slice))
@@ -237,23 +312,20 @@ func castBufferToSlice(buf unsafe.Pointer, length int) []byte {
     return slice
 }
 
-func castBuffer(javaOutputBuf *C.char, length int) []byte {
-    bufSize := length
-    if bufSize < EIP196PreallocateForResult {
+func castBufferEIP196(javaOutputBuf *C.char, length *int) []byte {
+    bufSize := *length
+    if bufSize != EIP196PreallocateForResult {
       bufSize = EIP196PreallocateForResult
     }
     return (*[EIP196PreallocateForResult]byte)(unsafe.Pointer(javaOutputBuf))[:bufSize:bufSize]
 }
 
-func castErrorBuffer(javaOutputBuf *C.char, length int) []byte {
-    bufSize := length
-    if bufSize < EIP196PreallocateForError {
+func castErrorBufferEIP196(javaOutputBuf *C.char, length *int) []byte {
+    bufSize := *length
+    if bufSize != EIP196PreallocateForError {
       bufSize = EIP196PreallocateForError
     }
     return (*[EIP196PreallocateForError]byte)(unsafe.Pointer(javaOutputBuf))[:bufSize:bufSize]
-}
-
-func main() {
 }
 
 // generate g1Add test data suitable for unit test input csv
@@ -311,12 +383,5 @@ func Uint256ToStringBigEndian(number *big.Int) string {
 	return hex.EncodeToString(bytes)
 }
 
-// isAllZero checks if all elements in the byte slice are zero
-func isAllZero(data []byte) bool {
-    for _, b := range data {
-        if b != 0 {
-            return false
-        }
-    }
-    return true
+func main() {
 }