+++ gnark/src/main/java/org/hyperledger/besu/nativelib/gnark/LibGnarkEIP196.java
@@ -2,7 +2,6 @@ package org.hyperledger.besu.nativelib.gnark;
 
 import com.sun.jna.Native;
 import com.sun.jna.ptr.IntByReference;
-import org.apache.tuweni.bytes.Bytes;
 
 public class LibGnarkEIP196 {
 
@@ -42,50 +41,36 @@ public class LibGnarkEIP196 {
     int ret = -1;
     switch(op) {
       case EIP196_ADD_OPERATION_RAW_VALUE:
-        ret = eip196altbn128G1Add(i, output, err, i_len,
-            EIP196_PREALLOCATE_FOR_RESULT_BYTES, EIP196_PREALLOCATE_FOR_ERROR_BYTES);
-        o_len.setValue(64);
+        ret = eip196altbn128G1Add(i, output, err, i_len, o_len, err_len);
         break;
       case  EIP196_MUL_OPERATION_RAW_VALUE:
-        ret = eip196altbn128G1Mul(i, output, err, i_len,
-            EIP196_PREALLOCATE_FOR_RESULT_BYTES, EIP196_PREALLOCATE_FOR_ERROR_BYTES);
-        o_len.setValue(64);
+        ret = eip196altbn128G1Mul(i, output, err, i_len, o_len, err_len);
         break;
       case EIP196_PAIR_OPERATION_RAW_VALUE:
-        ret = eip196altbn128Pairing(i, output, err, i_len,
-            EIP196_PREALLOCATE_FOR_RESULT_BYTES, EIP196_PREALLOCATE_FOR_ERROR_BYTES);
-        o_len.setValue(32);
+        ret = eip196altbn128Pairing(i, output, err, i_len, o_len, err_len);
         break;
       default:
         throw new RuntimeException("Not Implemented EIP-196 operation " + op);
     }
 
-    if (ret != 0) {
-      var errBytes = Bytes.wrap(err);
-      err_len.setValue(errBytes.size() - errBytes.numberOfTrailingZeroBytes());
-      o_len.setValue(0);
-    } else {
-      err_len.setValue(0);
-    }
     return ret;
   }
-
 
   public static native int eip196altbn128G1Add(
       byte[] input,
       byte[] output,
       byte[] error,
-      int inputSize, int outputSize, int err_len);
+      int inputSize, IntByReference outputSize, IntByReference err_len);
 
   public static native int eip196altbn128G1Mul(
       byte[] input,
       byte[] output,
       byte[] error,
-      int inputSize, int output_len, int err_len);
+      int inputSize, IntByReference output_len, IntByReference err_len);
 
   public static native int eip196altbn128Pairing(
       byte[] input,
       byte[] output,
       byte[] error,
-      int inputSize, int output_len, int err_len);
+      int inputSize, IntByReference output_len, IntByReference err_len);
 }