+++ gnark/src/main/java/org/hyperledger/besu/nativelib/gnark/LibGnarkEIP2537.java
@@ -3,9 +3,6 @@ package org.hyperledger.besu.nativelib.gnark;
 import com.sun.jna.Library;
 import com.sun.jna.Native;
 import com.sun.jna.ptr.IntByReference;
-import org.apache.tuweni.bytes.Bytes;
-
-import java.nio.charset.StandardCharsets;
 
 public class LibGnarkEIP2537 implements Library {
 
@@ -110,8 +107,7 @@ public class LibGnarkEIP2537 implements Library {
     }
 
     if (ret != 0) {
-      var errBytes = Bytes.wrap(err);
-      err_len.setValue(errBytes.size() - errBytes.numberOfTrailingZeroBytes());
+      err_len.setValue(LibGnarkUtils.findFirstTrailingZeroIndex(err));
       o_len.setValue(0);
     } else {
       err_len.setValue(0);