+++ gnark/src/main/java/org/hyperledger/besu/nativelib/gnark/LibGnarkUtils.java
@@ -1,0 +1,14 @@
+package org.hyperledger.besu.nativelib.gnark;
+
+public class LibGnarkUtils {
+
+  public static int findFirstTrailingZeroIndex(byte[] array) {
+    for (int i = array.length - 1; i >= 0; i--) {
+      if (array[i] != 0) {
+        return i + 1; // The first trailing zero is after this non-zero byte
+      }
+    }
+    // If all bytes are zero, return 0
+    return 0;
+  }
+}