+++ gnark/src/test/java/org/hyperledger/besu/nativelib/gnark/BLS12G1MultiExpPrecompiledContractTest.java
@@ -77,6 +77,7 @@ public class BLS12G1MultiExpPrecompiledContractTest {
     } else {
       final Bytes actualComputation = Bytes.wrap(output, 0, outputLength.getValue());
       assertThat(actualComputation).isEqualTo(expectedComputation);
+      assertThat(notes).isEmpty();
     }
   }
 }