+++ gnark/src/test/java/org/hyperledger/besu/nativelib/gnark/BLS12G2AddPrecompiledContractTest.java
@@ -82,6 +82,7 @@ public class BLS12G2AddPrecompiledContractTest {
     } else {
       final Bytes actualComputation = Bytes.wrap(output, 0, outputLength.getValue());
       assertThat(actualComputation).isEqualTo(expectedComputation);
+      assertThat(notes).isEmpty();
     }
   }
 }